Use increment permutation in hash-to-RSA
I did some investigations into the increment permutation as a PRG of
sorts for RSA generation. The idea is this: Normally, to generate a
2048b random number, you need to do a sponge or something with 2048b of
output. However, this is expensive.

Another idea is to use a compression function to produce ~250b of
output, and then use a permutation to extend it to the necessary 2048b.
The question then is: what permutation?

Let H be the initial hash output. The trivial permutation would produce
a number like:
   H H H H ... H

which is divisible by:
   1 1 1 1 ... 1

So that doesn't work. However, I spent some time investigating the +1
permutation, and it does seem to work. Specifically, a pool of number
generated with the +1 permutation does not seem to have any more common
factors that one generated truly at random.