Add high bit to Pocklington entropy
Since the pocklington hash generates a challenge which becomes a modulus
throughout the codebase, making sure that it has high bits decreases a
*whole* bunch of quotient sizes by the width of the challenge.

By adding a high bit to pocklington entropy, we increase the challenge
size by a the number of pocklington extensions (3 for 128b entropy), but
the shrink quotients by the length of the challenge (~180b) -- which is
a clear win.

Reduces the constaint count from 2217 -> 1819 per transaction.

Starting from 4476 constraints/set-op, we've shrunk by -59% in total.